Remove insecure configurations
In order to protect both IXP operator and IXP participants, we should
ensure that using arouteserver never leads to an insecure mode of
operation.

While arouteserver should offer a degree on flexibility in terms of
fine tuning what IRRs are used, whether ARIN WHOIS or RPKI is used
or not - there should not be a configuration path that leads to no
filters at all. Offering fully insecure modes is not user friendly.

This fixes #22