Fix: backup of the wrong custom template file
Scenario: a template file is locally edited to implement some
custom changes. After an upgrade, 'arouteserver setup-templates'
is executed and the program shows the following prompt:

   the installed version of templates/XXXX/YYYY.j2 is not
   aligned with the one used by the current version of the
   program; moreover, it seems that it has been edited after
   the installation on the local system

   Do you want to create a backup copy into <path>? [YES/no]

Due to this bug, the tool was creating a copy of the new file,
instead of the locally edited one, thus making the backup file not
useful at all.