Fix: OpenBGPD prepend comms deletion for 32bit ASNs
A general.yml configuration on which prepend communities were
configured with only 'std' commmunities...

    prepend_once_to_any:
      std: "65501:65501"

... led to a broken OpenBGPD configuration for the outbound
policy towards 32bit ASNs:

    # prepend_once_to_any; remove further "prepend" communities from matching routes
    match to 192.0.2.1 community 65501:65501 set prepend-neighbor 1
    match to 192.0.2.1 community 65501:65501 set {  }

The error was on the line of configuration where all the prepend
communities are removed from routes that already matched a prepend
community. The resulting 'set {}' statement on that line was empty
because the 32bit ASN for which the outbound policy was generated
was preventing the 'std' community from being matched, even though
it shouldn't matter for 'std' communities not having a 'peer_as'
part.

Fix #98.