OpenBGPD: refactor prepending actions
To simplify the config generated around the prepending communities
the implementation of each action now depends on the presence of
an internal community, INTCOMM_PROCESS_PREPEND_COMMS, that is
initially attached to the routes, and that is dropped as soon as
one prepending community is matched.
In this way we can avoid removing all the prepending communities
every time the first one is matched, while keeping the same
priority order in which they are processed.

This is related to #97 and also improves #98.