Update dependencies (requirements.txt)
Fix #106

This is to have requirements.txt including the version of the packages
that we actually need. Specifically, `indent` was changed in Jinja2 3.0
to also accept the character to be used for the indentantion (which is
what we're passing today in the templates); previous versions are not
able to deal with it.