Fix: `--merge-from-custom-file` overriding configs
The `--merge-from-custom-file` option of the `clients-from-euroix`
command is supposed to merge the configurations from a local file
into the final content that it generates for clients.yml.
However, a bug was triggering an undesired behaviour, for which the
`cfg` settings from the local file were overriding those automatically
generated by the command. This was leading to the removal of information
such as max-prefix and as-set that were originally populated using the
Euro-IX records.