Modify ADD-PATH for OpenBGPD 8.7
ADD-PATH in OpenBGPD 8.7 seems to work in a different way with respect
to previous releases. When the `send best` option is configured, the
best path is not advertised, but only the other paths are.
This is currently being investigated by the OpenBGPD developers.
The change from this commit, however, seems to be working fine to
restore the correct behaviour from a high-level perspective.